+++ docs/content/migrate/v3.md
@@ -689,13 +689,13 @@ Since `v3.7.0`, the `Host` and `HostSNI` matchers support wildcard subdomain mat
 This allows matching any direct subdomain of a domain with a single-level wildcard prefix.
 For example, `*.example.com` matches `foo.example.com` but not `foo.bar.example.com` or `example.com` itself.
 
-This feature is only available with the v3 rule syntax (the default). It is not supported with the deprecated v2 rule syntax.
+This feature is only available with the v3 rule syntax (the default).
 
 #### TLSOptions with Wildcard Domains
 
 Since `v3.7.0`, TLSOptions can now be associated with routers using wildcard `Host` and `HostSNI` matchers (e.g., `Host(`*.example.com`)`).
 This enables configuring different TLS options for wildcard domains.
 
-Previously, TLSOptions selection was limited to exact `Host` matches, and using `HostRegexp` or wildcards would fall back to the default TLS options with a warning message like: `No domain found in rule HostRegexp(...) the TLS options applied for this router will depend on the SNI of each request`.
+Previously, TLSOptions selection was limited to exact `Host` matches, and using `HostRegexp` or wildcards would fall back to the default TLS options with a warning message like: `No domain found in rule HostRegexp(...) the TLS option foo cannot be applied`.
 
 Note: TLSOptions for `HostRegexp` matchers remains unsupported. Use wildcard `Host` matchers as an alternative.
