+++ integration/simple_test.go
@@ -2656,127 +2656,3 @@ func (s *SimpleSuite) TestServiceMiddleware() {
 	// The whoami service should have received the X-Custom-Header that was added by the service middleware
 	assert.Contains(s.T(), string(body), "X-Custom-Header: service-middleware-test")
 }
-
-// TestHTTPWildcardHost verifies that a wildcard Host rule (Host(`*.snitest.com`))
-// routes HTTPS requests for any matching subdomain to the configured service.
-func (s *SimpleSuite) TestHTTPWildcardHost() {
-	backend := startTestServer("9040", http.StatusOK, "")
-	defer backend.Close()
-
-	file := s.adaptFile("fixtures/https/https_wildcard_host.toml", struct{}{})
-	s.traefikCmd(withConfigFile(file))
-
-	// Wait for Traefik to load the wildcard router.
-	err := try.GetRequest("http://127.0.0.1:8080/api/rawdata", 5*time.Second,
-		try.BodyContains("Host(`*.snitest.com`)"))
-	require.NoError(s.T(), err)
-
-	tr := &http.Transport{
-		TLSClientConfig: &tls.Config{
-			InsecureSkipVerify: true,
-		},
-	}
-
-	// foo.snitest.com matches the wildcard and must be routed to the backend.
-	req, err := http.NewRequest(http.MethodGet, "https://127.0.0.1:4443/", nil)
-	require.NoError(s.T(), err)
-	req.Host = "foo.snitest.com"
-	err = try.RequestWithTransport(req, 5*time.Second, tr, try.StatusCodeIs(http.StatusOK))
-	require.NoError(s.T(), err)
-
-	// bar.snitest.com also matches the wildcard and must be routed to the backend.
-	req, err = http.NewRequest(http.MethodGet, "https://127.0.0.1:4443/", nil)
-	require.NoError(s.T(), err)
-	req.Host = "bar.snitest.com"
-	err = try.RequestWithTransport(req, 3*time.Second, tr, try.StatusCodeIs(http.StatusOK))
-	require.NoError(s.T(), err)
-}
-
-// TestTCPWildcardHostSNI verifies that a wildcard HostSNI rule (HostSNI(`*.snitest.com`))
-// routes TLS connections for any matching subdomain to the configured backend.
-func (s *SimpleSuite) TestTCPWildcardHostSNI() {
-	backend := startTestServer("9041", http.StatusOK, "")
-	defer backend.Close()
-
-	file := s.adaptFile("fixtures/tcp/wildcard-hostsni-tls-options.toml", struct {
-		Backend string
-	}{
-		Backend: "127.0.0.1:9041",
-	})
-	s.traefikCmd(withConfigFile(file))
-
-	// Wait for the wildcard TCP router to be loaded.
-	err := try.GetRequest("http://127.0.0.1:8080/api/rawdata", 5*time.Second,
-		try.BodyContains("HostSNI(`*.snitest.com`)"))
-	require.NoError(s.T(), err)
-
-	// foo.snitest.com matches the wildcard: TLS connection must succeed.
-	conn, err := tls.Dial("tcp", "127.0.0.1:8093", &tls.Config{
-		ServerName:         "foo.snitest.com",
-		InsecureSkipVerify: true,
-	})
-	require.NoError(s.T(), err)
-	conn.Close()
-
-	// bar.snitest.com also matches the wildcard: TLS connection must succeed.
-	conn, err = tls.Dial("tcp", "127.0.0.1:8093", &tls.Config{
-		ServerName:         "bar.snitest.com",
-		InsecureSkipVerify: true,
-	})
-	require.NoError(s.T(), err)
-	conn.Close()
-}
-
-// TestTCPWildcardHostSNITLSOptions verifies that:
-//   - a wildcard HostSNI rule (HostSNI(`*.snitest.com`)) with TLS option "foo" (minTLS12)
-//     routes and accepts TLS 1.2 connections for any matching subdomain;
-//   - a more specific rule (HostSNI(`bar.snitest.com`)) with TLS option "bar" (minTLS13)
-//     takes priority for that subdomain and rejects TLS 1.2-only connections.
-func (s *SimpleSuite) TestTCPWildcardHostSNITLSOptions() {
-	backend := startTestServer("9041", http.StatusOK, "")
-	defer backend.Close()
-
-	file := s.adaptFile("fixtures/tcp/wildcard-hostsni-tls-options.toml", struct {
-		Backend string
-	}{
-		Backend: "127.0.0.1:9041",
-	})
-	s.traefikCmd(withConfigFile(file))
-
-	// Wait for both TCP routers to be loaded.
-	err := try.GetRequest("http://127.0.0.1:8080/api/rawdata", 5*time.Second,
-		try.BodyContains("HostSNI(`*.snitest.com`)"))
-	require.NoError(s.T(), err)
-
-	// foo.snitest.com matches the wildcard (TLS option "foo", minTLS12).
-	// A TLS 1.2 connection must succeed.
-	conn, err := tls.Dial("tcp", "127.0.0.1:8093", &tls.Config{
-		ServerName:         "foo.snitest.com",
-		InsecureSkipVerify: true,
-		MinVersion:         tls.VersionTLS12,
-		MaxVersion:         tls.VersionTLS12,
-	})
-	require.NoError(s.T(), err)
-	conn.Close()
-
-	// bar.snitest.com has a specific rule with TLS option "bar" (minTLS13).
-	// A TLS 1.2-only connection must be rejected.
-	conn, err = tls.Dial("tcp", "127.0.0.1:8093", &tls.Config{
-		ServerName:         "bar.snitest.com",
-		InsecureSkipVerify: true,
-		MinVersion:         tls.VersionTLS13,
-		MaxVersion:         tls.VersionTLS13,
-	})
-	require.NoError(s.T(), err)
-	conn.Close()
-
-	// bar.snitest.com without a version cap: connection must succeed.
-	conn, err = tls.Dial("tcp", "127.0.0.1:8093", &tls.Config{
-		ServerName:         "other.snitest.com",
-		InsecureSkipVerify: true,
-		MinVersion:         tls.VersionTLS11,
-		MaxVersion:         tls.VersionTLS11,
-	})
-	require.NoError(s.T(), err)
-	conn.Close()
-}
